ReapplyRules: removing explicit removal of ReapplyRules flag
This was originally intended to be a one-shot option to reapply rules
once and only once for a window, but there is a use case for recurring
appliance of rules.

There may need to be an exception introduced at some point. For example
if the same rule is continuously applied then there may need to be a
limit of sorts.

The ReapplyRules flag is still experimental.