Use wmcheckwin as the barrier between tiled and floating windows
Instead of using the bar as a reference (in a bare dwm tiled windows
are placed below the bar and floating windows are placed above the
bar), use wmcheckwin for this purpose.

The reasoning is that a monitor can have more than one bar and the
order of the bars is not guaranteed, so if a bar is removed and
another takes its place (as in the case of substituting an internal
bar with an external one), then the tiled vs floating reference can
shift which can cause tiled windows to be place above floating
windows.