Major workspace refactoring
Most of these changes are non-functional refactoring to shorten some
code and to generally make the workspace implementation more readable
and maintainable.

There are a few bugfixes as well that was found during refactoring and
review.

Main feature added:

There was an edge case where if all workspaces were pinned to, say,
two monitors and a third one were added to the mix - and that third
monitor happened to end up as the primary monitor, then this could
lead to a crash due to having a null selected workspace.

Fixing that one crash would still leave the user with a monitor that
has no workspaces and they would not be able to open a window there
for example.

While the situation is something that can be "explained" by that
there are no workspaces available on the monitor it is confusing
and unintuitive behaviour. To address this a fallback "null"
workspace was implemented that will be created automatically if the
monitor ends up with no workspaces. This will allow the user to spawn
programs and do everything as normal, while not having any of the
defined (via wsrules) workspaces on the monitor.

These "null" workspaces intentionally do not have a workspace icon.

If a monitor has a "null" workspace with two clients on it, and one
of the other workspaces are moved to the monitor, then the "null"
workspace will disappear and the two clients will be moved over to
the workspace that was moved over.

In practice this solution should in principle guarantee that there
is always a selected workspace on all monitors (i.e. selws and
mon->selws should never be NULL).