UnmapNotify: addressing edge cases
The revised unmapnotify event handling, due to now receiving
notificiations when unmapping windows in relation to hiding them,
worked in general fine, but it had some edge cases.

The primary issue is that it would miss legitimate unmap events
like when pressing Esc to close the GtkFileChooserDialog window,
in which case the window would legitimately unmap, but would not
be destroyed.

This would leave empty clients that could not be shown and could
not be closed.

Another strategy is to flag when one expect to receive an unmap
notification for a window, and fend off the unmapnotify event as
and when that comes through. This naturally involves clearing said
flag so that a legitimate unmap notification can be handled
correctly. Attempting this quickly revealed that for every unmap
action we receive two unmap notifications. Checking the serial
value of the event revealed that it was indeed a duplicate event.

The reason for the duplicate event is because the window manager
uses a SubstructureRedirectMask on the root window to intercpet
window state changes, which makes it receive UnmapNotify events
for windows it manages. If both the window manager and the X
server generate UnmapNotify events then there will be duplicates.

A strategy to de-duplicate such notifications was to keep a record
of the last event serial processed and skip events if a duplicate
is received. On that topic my first approach was to de-duplicate
events in the structurenotify function before it reaches the
unmapnotify, destroynotify or configurenotify functions. What I
learnt was that you can have two UnmapNotify events with the same
serial, but for two different windows. This can happen when an
application maintains more than one window and there is a cascading
collapse when the parent window dies. So filtering on the serial
and the window ID is necessary. More so, in such a situation we
can receive two legitimate notifications where the serial and
window IDs are the same, but the event type is different. As such
it became clear that the de-duplication need to happen in the
unmapnotify and destroynotify functions rather than in
structurenotify. De-duplication was skipped for the configure
notifications because there we are only interested in changes for
the root window.

A secondary issue was that the show and hide functions can be
called while the window is already shown or already hidden, which
could screw with the information that we rely on to know whether
to expect an unmap notification to come through for a window.

One strategy is to grab XGetWindowAttributes and check the map
state of the window before deciding whether to trigger the show
and hide logic. Minor performance impact. In the end this approach
did not add anything over simply tracking the state in a new
client variable c->shown.

A third issue was that when spamming workspace switches via a
script for example events would stack up and at the end there
would be a flow of unmapnotify events coming through. The solution
for that was to increment an expecting unmap event counter for
every hide call, and decrement that counter every time an unmap
notification is received for the window.

Summary of the three changes introduced:
   - only one show call will be handled following a hide call,
     and vice-versa
   - the UnmapNotify and DestroyNotify events will now be
     de-duplicated in their respective functions
   - hiding a window will increment an expecting unmap counter,
     which will be decremented when unmap notifications come
     through for the window