swallow: leave swallow stack with unswallowed window
When having multiple clients being swallowed in a stack the
remaining clients in the stack would remain with the topmost client.

The reasoning beind this was that if the user decide to manually
unswallow a window then it makes sense for the focus to remain on
the original client, and it makes sense that one can continue
hitting the same keybinding to unswallow more.

This does not work that well if an unswallow is forced in unmanage,
in which case it would take the remaining swallow stack with it to
the grave.

So to make this behave more or less the same we now leave the
swallow stack with the unswallowed window and we give focus to the
unswallowed window.