vanitygaps: removing gapp prefix from internal variables
but leaving the configuration values as-is. The intention is to make
the code more readable and that variables ih, iv, oh and ov are in
relation to gaps should hopefully be clear without the "gap pixel"
prefix.

For the configuration the extra clarity is presumed to be helpful
and changes would break existing configuration. If changes are to
be made in the future then likely longer more descriptive variables
would be more appropriate.