aspectgrid: refinements
In order to avoid scenarios where the last client ends up with a lot
of space that could have been better used by the two prior windows
the aspectgrid algorithm will now try to include the last client as
well to see if that arrangement is a better fit.

Also the border width should not be taken into account when doing
calculations using the aspect ratio. This has to do with that the
border is not actually part of the window size as such and should
therefore not count when applying the aspect ratio in calculations.