Do not restore window border on unmanage.
The original code was added back in April 2007 for dwm with commit
ad2508f9578195e703033df3cceb441f716261bd and comment:
   | fixed the border issue for mplayer, ff is definately broken when using F11 (fullscreen mode)

The changes were primarily in the manage function, and no explicit reason was given for restoring
the original border when unmanaging the window.

I am removing that code because when a restart of dusk is made then the border is restored when
dusk is exiting/restarting (as all windows are unmanged and not destroyed) and the border is
removed again when dusk starts up again and beings managing the window again. This commit is to
remove that border flicker.

The removal of the Locked flag is to allow the border to be removed following a restart.