Reverting: set _NET_WM_STATE_HIDDEN when hiding windows
The _NET_WM_STATE_HIDDEN net wm state happens to only be set in
gnome when a window is minimised.

When changing desktops / workspaces the only property that changes
is the basic wm state going from NormalState to IconicState.

As such the conclusion is that the change of viewport results in
the window becoming "out of view", which is what discord / electron
may be using to determine that the window is not visible on screen
and makes it send notifications also for the current channel.