Fix for Xephyr disappearing
Some programs like Xephyr keep sending configure requests with their
current position. If the window is not shown at the time when dusk
restarts then the Xephyr window would not be present on the designated
workspace the first time the user navigated to it. This has to do with
that the window needed a resize, placing it instead at its present
negative position. The fix allows the window size to change in this
scenario, but the window should still be moved into view when moving
to the designated workspace.