swallow: retain first level swallow across restarts
When restarting one does not have control over the order in which
existing windows are returned when querying / scanning for windows.

As such there may be a terminal window that has been swallowed by
e.g. mpv. If the terminal is mapped first then the mpv window is
swallowed as normal when that maps, but if the mpv window is mapped
first then the terminal would open separately.

These changes allow for the terminal window to take the place of
the mpv window and let the mpv window swallow the terminal window
retrospectively.

If a chain of windows have been manually swallowed then these are
expected to untangle following a restart.

If there is a process dependency in a chain of windows, e.g.
starting a terminal from a terminal from a terminal, then these
may still stack following a restart - but the order is not
guaranteed.