drawws: call drawbars before arrange
If bars are set up such that they appear and disappear depending
on whether there are floating windows, for example, then moving
from a workspace having two bars to a workspace that only has one
would result in an initial arrange (through draws) where two bars
are taken into account, and then when drawbars was called it would
work out that one bar was to be removed, then the workspace areas
would be updated and another arrange would be triggered.

By calling drawbars first we avoid the initial arrange taking the
extra bar into account when it shouldn't.