Refactoring: Replacing XImage logic with XRender for performance
The previous commit functionally allowed us to capture and scale
workspace screenshots while dropping the imlib2 dependency.

Doing a performance comparison, scaling a resolution of 5120x1440
down to a quarter (pfact = 0.25) giving a preview resolution of
1280x360 my findings were that on average it would take 0.13235s
(132ms) to create the preview pixmap.

Comparatively the original imlib2 implementation took on average
0.02963s (29.63ms).

This suggests that the new solution has a 4x to 5x performance
penalty.

The imlib2 implementation may likely be faster because of the use
of shared memory reducing the need to transfer large amounts of
data between the X server and the client.

Further analysis found that it is possible to avoid the transfer
of data altogether by letting XRender capture and scale the
screenshot on the server side, while offloading the calculations
to the GPU.

Performance analysis found that this method took on average
0.0000514s (0.0514ms) on the same setup, suggesting a significant
boost in performance.

This led to some further refactoring to make the window icons and
status2d image handling use the same picture creation and scaling
functions.